Unify language used in expectations and their reports.
Prior to this commit, ToExecuteCommandMatching() and ToRecordEventMatching()
described their failures using slightly different wording ('satisfies' vs
'matches') for example.

This changes the failure reports to use language consistent with the names of
the expectation functions. It also adds the file/line of the predication
function to aid in locating which actual predicate failed to match.